move relation and actor level metrics away from cluster_essential
